User Response page and update to routes to only allow admins and logged in users
Changed some routes to admin and logged-in users only.
Created a page for users to respond to jobs that have been offered

+++ routes/web.php
@@ -19,21 +19,23 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/jobs', [JobController::class, 'index'])->name(('jobs.index'));
-Route::get('/jobs/create', [JobController::class,'create'])->name(('jobs.create'))->middleware(('auth'));
-Route::post('/jobs', [JobController::class,'store'])->name(('jobs.store'))->middleware(('auth'));
-Route::get('/jobs/{id}', [JobController::class,'show'])->name(('jobs.show'));
-Route::delete('/jobs/{id}', [JobController::class,'destroy'])->name(('jobs.destroy'))->middleware(('auth'));
+Route::get('/jobs', [JobController::class, 'index'])->name(('jobs.index'))->middleware(('auth'));
+Route::get('/jobs/create', [JobController::class,'create'])->name(('jobs.create'))->middleware(['auth','auth.isAdmin']);
+Route::post('/jobs', [JobController::class,'store'])->name(('jobs.store'))->middleware(['auth','auth.isAdmin']);
+Route::get('/jobs/{id}', [JobController::class,'show'])->name(('jobs.show'))->middleware(('auth'));
+Route::delete('/jobs/{id}', [JobController::class,'destroy'])->name(('jobs.destroy'))->middleware(['auth','auth.isAdmin']);
 
 
-Route::get('/applications', [ApplicationsController::class, 'index'])->name(('applications.index'))->middleware(('auth'));
+Route::get('/applications', [ApplicationsController::class, 'index'])->name(('applications.index'))->middleware(['auth','auth.isAdmin']);
 Route::get('/applications/myApplications', [ApplicationsController::class, 'userShow'])->name(('applications.userShow'))->middleware(('auth'));
 
 
 Route::get('/applications/create', [ApplicationsController::class,'create'])->name(('applications.create'))->middleware(('auth'));
 Route::post('/applications', [ApplicationsController::class,'store'])->name(('applications.store'))->middleware(('auth'));
-Route::get('/applications/{id}', [ApplicationsController::class,'show'])->name(('applications.show'));
-Route::get('/applications/edit/{id}', [ApplicationsController::class, 'edit'])->name(('applications.edit'));
+Route::get('/applications/{id}', [ApplicationsController::class,'show'])->name(('applications.show'))->middleware(['auth','auth.isAdmin']);
+Route::get('/applications/edit/{id}', [ApplicationsController::class, 'edit'])->name(('applications.edit'))->middleware(['auth','auth.isAdmin']);
+Route::get('/applications/userResponse/{id}', [ApplicationsController::class, 'userResponse'])->name(('applications.userResponse'));
+
 Route::patch('/applications/edit/{id}', [ApplicationsController::class, 'update'])->name(('applications.update'));
